+++ covid/__main__.py
@@ -1,3 +1,3 @@
-from .core import hmm
+from .core import start
 
-hmm()
+start()